DERP-629: Improve check for NVM
The previous check for NVM was failing due to the fact that the
nvm script returns an exit code of 3 when sourced, which was
preventing the `nvm install` and `nvm use` commands from executing.

Instead, we attempt to source the `nvm.sh` script if the nvm command
does not exist, and then we run `nvm install` and `nvm use` regardless.

Note that even if these commands fail, that doesn't mean that we should
not proceed. There is an existing check for Node version, which will
determine whether the script should continue.

https://london.atlassian.net/browse/DERP-629

[ Partial fix for DERP-629 ]